Initial Commit
master contains a basic test script that currently works with the hardware (SIM7600SA + Arduino Mega 2560). 

example_master contains a script pulled from the internet with full functionality, however, this does not seem to be a plug-and-play solution. The goal is to unit test/unit implement example_master to master.

broken_gps mirrors master but with a different GPS function pulled from a different source (lost now).